[commit] add datax job

+++ xxl-job-admin/src/main/java/com/xxl/job/admin/core/util/LocalCacheUtil.java
@@ -1,5 +1,7 @@
 package com.xxl.job.admin.core.util;
 
+import java.util.List;
+import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
@@ -11,7 +13,8 @@ import java.util.concurrent.ConcurrentMap;
 public class LocalCacheUtil {
 
     private static ConcurrentMap<String, LocalCacheData> cacheRepository = new ConcurrentHashMap<String, LocalCacheData>();   // 类型建议用抽象父类，兼容性更好；
-    private static class LocalCacheData{
+
+    private static class LocalCacheData {
         private String key;
         private Object val;
         private long timeoutTime;
@@ -59,13 +62,13 @@ public class LocalCacheUtil {
      * @param cacheTime
      * @return
      */
-    public static boolean set(String key, Object val, long cacheTime){
+    public static boolean set(String key, Object val, long cacheTime) {
 
         // clean timeout cache, before set new cache (avoid cache too much)
         cleanTimeoutCache();
 
         // set new cache
-        if (key==null || key.trim().length()==0) {
+        if (key == null || key.trim().length() == 0) {
             return false;
         }
         if (val == null) {
@@ -81,16 +84,27 @@ public class LocalCacheUtil {
     }
 
     /**
+     *
+     * @return
+     */
+    public static Set<String> list() {
+        return cacheRepository.keySet();
+    }
+
+    /**
      * remove cache
      *
      * @param key
      * @return
      */
-    public static boolean remove(String key){
-        if (key==null || key.trim().length()==0) {
+    public static boolean remove(String key) {
+        if (key == null || key.trim().length() == 0) {
             return false;
         }
-        cacheRepository.remove(key);
+        if (cacheRepository.containsKey(key)) {
+            cacheRepository.remove(key);
+        }
+
         return true;
     }
 
@@ -100,12 +114,12 @@ public class LocalCacheUtil {
      * @param key
      * @return
      */
-    public static Object get(String key){
-        if (key==null || key.trim().length()==0) {
+    public static Object get(String key) {
+        if (key == null || key.trim().length() == 0) {
             return null;
         }
         LocalCacheData localCacheData = cacheRepository.get(key);
-        if (localCacheData!=null && System.currentTimeMillis()<localCacheData.getTimeoutTime()) {
+        if (localCacheData != null && System.currentTimeMillis() < localCacheData.getTimeoutTime()) {
             return localCacheData.getVal();
         } else {
             remove(key);
@@ -118,11 +132,11 @@ public class LocalCacheUtil {
      *
      * @return
      */
-    public static boolean cleanTimeoutCache(){
+    public static boolean cleanTimeoutCache() {
         if (!cacheRepository.keySet().isEmpty()) {
-            for (String key: cacheRepository.keySet()) {
+            for (String key : cacheRepository.keySet()) {
                 LocalCacheData localCacheData = cacheRepository.get(key);
-                if (localCacheData!=null && System.currentTimeMillis()>=localCacheData.getTimeoutTime()) {
+                if (localCacheData != null && System.currentTimeMillis() >= localCacheData.getTimeoutTime()) {
                     cacheRepository.remove(key);
                 }
             }
